modify: CommentController JWT 수정

+++ src/main/java/com/ssafy/fit/controller/CommentController.java
@@ -2,8 +2,10 @@ package com.ssafy.fit.controller;
 
 import com.ssafy.fit.model.dto.Comment;
 import com.ssafy.fit.model.service.CommentService;
+import com.ssafy.fit.util.JwtUtil;
 import io.swagger.annotations.Api;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -15,6 +17,8 @@ import java.util.List;
 @Api(tags = "댓글 컨트롤러")
 public class CommentController {
 
+    @Autowired
+    private JwtUtil jwtUtil;
     @Autowired
     private CommentService commentService;
 
@@ -36,24 +40,49 @@ public class CommentController {
 
     // 댓글 작성
     @PostMapping("/{articleNo}/comment")
-    public ResponseEntity<String> insertComment(@PathVariable int articleNo, Comment comment){
+    public ResponseEntity<String> insertComment(@PathVariable int articleNo, Comment comment, @RequestHeader HttpHeaders header) throws Exception{
+        String token = header.get("access-token").toString();
+        int requestUserNo = jwtUtil.getUserNoAtToken(token);
+
         comment.setArticleNo(articleNo);
+        comment.setUserNo(requestUserNo);
+
         commentService.insertComment(comment);
         return new ResponseEntity<String>("Insert Complete !", HttpStatus.CREATED);
     }
 
-    // 리뷰 삭제
+    // 댓글 삭제
     @DeleteMapping("/comment/{commentNo}")
-    public ResponseEntity<String> deleteComment(@PathVariable int commentNo){
-        commentService.deleteComment(commentNo);
-        return new ResponseEntity<String>("Delete Complete!", HttpStatus.OK);
+    public ResponseEntity<String> deleteComment(@PathVariable int commentNo, @RequestHeader HttpHeaders header) throws Exception{
+        String token = header.get("access-token").toString();
+        int requestUserNo = jwtUtil.getUserNoAtToken(token);
+
+       Comment requestComment = commentService.selectCommentByNo(commentNo);
+
+        if(requestComment.getUserNo() == requestUserNo) {
+            commentService.deleteComment(commentNo);
+            return new ResponseEntity<String>("Delete Complete!", HttpStatus.OK);
+        }
+        else{
+            return new ResponseEntity<String>("No Permission to Delete",HttpStatus.BAD_REQUEST);
+        }
     }
 
-    // 리뷰 수정
+    // 댓글 수정
     @PutMapping("/comment/{commentNo}")
-    public ResponseEntity<String> updateComment(@PathVariable int commentNo, Comment comment){
-        commentService.updateComment(commentNo, comment.getContent());
-        return new ResponseEntity<String>("Update Complete!",HttpStatus.OK);
+    public ResponseEntity<String> updateComment(@PathVariable int commentNo, Comment comment, @RequestHeader HttpHeaders header) throws Exception{
+        String token = header.get("access-token").toString();
+        int requestUserNo = jwtUtil.getUserNoAtToken(token);
+
+        Comment requestComment = commentService.selectCommentByNo(commentNo);
+
+        if(requestComment.getUserNo() == requestUserNo) {
+            commentService.updateComment(commentNo, comment.getContent());
+            return new ResponseEntity<String>("Update Complete!", HttpStatus.OK);
+        }
+        else{
+            return new ResponseEntity<String>("No Permission to Modify",HttpStatus.BAD_REQUEST);
+        }
     }
 
 }
